Фиксация проекта 30.03/ Пройдены без ошибок все тесты Users и Films

+++ src/main/java/ru/yandex/practicum/filmorate/DAO/FilmDbStorage.java
@@ -39,28 +39,25 @@ public class FilmDbStorage implements FilmStorage {
         List<Film> films = jdbcTemplate.query(sqlFilmById, FilmDbStorage::makeFilm);
 
         for (Film film : films) {
-            //Создание списка жанров для конкретного film
             String sqlFilmGenres = "SELECT * " +
                     "FROM film_genres LEFT OUTER JOIN genres " +
                     "ON genres.genres_id = film_genres.genres_id " +
                     "WHERE film_id = ?";
             List<Genre> filmGenres = jdbcTemplate.query(
                     sqlFilmGenres, FilmDbStorage::makeGenre, film.getId());
-            //Установка жанров для film
             film.setGenres(filmGenres);
         }
-
-        //Вовзрат списка готовых film с жанрами
         return films;
     }
 
     @Override
     public Film createFilm(Film film) throws ValidationException, NotFoundException {
-
         Integer mpaId = film.getMpa().getId();
         String sqlMpa = "SELECT * FROM MPA_ratings WHERE rating_id = ?";
         List<MPA> mpaById = jdbcTemplate.query(sqlMpa, FilmDbStorage::makeMpa, mpaId);
-        validateExistMPA(mpaById, mpaId);
+        if (mpaById.isEmpty() || mpaById.get(0) == null || !mpaById.get(0).getId().equals(mpaId)) {
+            throw new NotFoundException(String.format("MPA with %s is not found", mpaId));
+        }
         MPA mpa = mpaById.get(0);
 
         Set<Integer> genreSet = film.getGenres()
@@ -74,14 +71,15 @@ public class FilmDbStorage implements FilmStorage {
             String sqlGenresId = "SELECT * FROM genres WHERE genres_id = ?";
             List<Genre> genresById = jdbcTemplate.query(
                     sqlGenresId, FilmDbStorage::makeGenre, id);
-            validateExistGenre(genresById, id);
+            if (genresById.isEmpty() || genresById.get(0) == null || !genresById.get(0).getId().equals(id)) {
+                throw new NotFoundException(String.format("Genre with %s is not found", id));
+            }
             filmGenres.add(genresById.get(0));
         }
 
         film.setGenres(filmGenres);
         film.setMpa(mpa);
 
-        // Добавление записи в таблицу films
         String sqlInsertFilm = "INSERT INTO films (film_name, description, release_date, duration, rating) " +
                 "VALUES (?, ?, ?, ?, ?)";
         KeyHolder keyHolder = new GeneratedKeyHolder();
@@ -96,11 +94,9 @@ public class FilmDbStorage implements FilmStorage {
             return ps;
         }, keyHolder);
 
-        // Получение id добавленного фильма
         int filmId = keyHolder.getKey().intValue();
         film.setId(filmId);
 
-        // Добавление записей в таблицу film_genres для каждого жанра фильма
         String sqlInsertFilmGenres = "INSERT INTO film_genres (film_id, genres_id) VALUES (?, ?)";
         for (Genre genre : film.getGenres()) {
             jdbcTemplate.update(sqlInsertFilmGenres, filmId, genre.getId());
@@ -109,42 +105,38 @@ public class FilmDbStorage implements FilmStorage {
     }
 
 
-
     @Override
     public Film updateFilm(Film film) throws NotFoundException, ValidationException {
         Integer mpaId = film.getMpa().getId();
         String sqlMpa = "SELECT * FROM MPA_ratings WHERE rating_id = ?";
         List<MPA> mpaById = jdbcTemplate.query(sqlMpa, FilmDbStorage::makeMpa, mpaId);
 
-validateExistMPA(mpaById, mpaId);
-
+        if (mpaById.isEmpty() || mpaById.get(0) == null || !mpaById.get(0).getId().equals(mpaId)) {
+            throw new NotFoundException(String.format("MPA with %s is not found", mpaId));
+        }
         MPA mpa = mpaById.get(0);
 
         Set<Integer> genreSet = film.getGenres()
                 .stream()
                 .map(Genre::getId)
                 .collect(Collectors.toSet());
-
-
         List<Genre> filmGenres = new ArrayList<>();
         for (Integer id : genreSet) {
             String sqlGenresId = "SELECT * FROM genres WHERE genres_id = ?";
             List<Genre> genresById = jdbcTemplate.query(
                     sqlGenresId, FilmDbStorage::makeGenre, id);
-
-validateExistGenre(genresById, id);
-
+            if (genresById.isEmpty() || genresById.get(0) == null || !genresById.get(0).getId().equals(id)) {
+                throw new NotFoundException(String.format("Genre with %s is not found", id));
+            }
             filmGenres.add(genresById.get(0));
         }
 
         film.setGenres(filmGenres);
         film.setMpa(mpa);
-               // Проверка существования фильма в базе данных
         if (getFilmById(film.getId()) == null) {
             throw new NotFoundException("Фильм не найден");
         }
 
-        // Обновление записи в таблице films
         String sqlUpdateFilm = "UPDATE films SET film_name = ?, description = ?, release_date = ?, duration = ?, rating = ? WHERE film_id = ?";
         jdbcTemplate.update(sqlUpdateFilm,
                 film.getName(),
@@ -154,11 +146,9 @@ validateExistGenre(genresById, id);
                 film.getMpa() != null ? film.getMpa().getId() : null,
                 film.getId());
 
-        // Удаление всех записей о жанрах фильма из таблицы film_genres
         String sqlDeleteFilmGenres = "DELETE FROM film_genres WHERE film_id = ?";
         jdbcTemplate.update(sqlDeleteFilmGenres, film.getId());
 
-        // Добавление записей о жанрах фильма в таблицу film_genres
         String sqlInsertFilmGenres = "INSERT INTO film_genres (film_id, genres_id) VALUES (?, ?)";
         for (Genre genre : film.getGenres()) {
             jdbcTemplate.update(sqlInsertFilmGenres, film.getId(), genre.getId());
@@ -167,53 +157,57 @@ validateExistGenre(genresById, id);
     }
 
 
-
     @Override
     public Film getFilmById(Integer id) {
-        return jdbcTemplate.query("SELECT * FROM films WHERE film_id = ?;", new Object[]{id}, new BeanPropertyRowMapper<>(Film.class))
-                .stream()
-                .findAny()
-                .orElseThrow(() -> new NotFoundException("Film not found with film_id: " + id));
+        String sql = "SELECT * FROM films " +
+                "LEFT OUTER JOIN mpa_ratings " +
+                "ON films.rating = mpa_ratings.rating_id " +
+                "WHERE film_id = ?";
+        List<Film> films = jdbcTemplate.query(sql, FilmDbStorage::makeFilm, id);
+        if (films.isEmpty()) {
+            return null;
+        }
+        Film film = films.get(0);
+
+        String sqlLikesCount = "SELECT COUNT(*) AS likes_count FROM likes WHERE film_id = ?";
+        List<Integer> likesCountList = jdbcTemplate.query(sqlLikesCount, (rs, rowNum) -> rs.getInt("likes_count"), id);
+        if (!likesCountList.isEmpty()) {
+            int likesCount = likesCountList.get(0);
+        }
+
+        String sqlGenres = "SELECT * FROM film_genres " +
+                "LEFT OUTER JOIN genres " +
+                "ON film_genres.genres_id = genres.genres_id " +
+                "WHERE film_id = ?";
+        List<Genre> genres = jdbcTemplate.query(sqlGenres, FilmDbStorage::makeGenre, id);
+        film.setGenres(genres);
+
+        return film;
     }
+
 
     @Override
     public Film installingLike(Integer filmId, Integer userId) {
-        //Валидация user_id
-        String sqlFilmLikesById = "SELECT * FROM likes WHERE FILM_ID = ?";
-        List<Integer> filmLikesList = jdbcTemplate.query(sqlFilmLikesById, (rs, rowNum) -> rs.getInt("USER_ID"), filmId);
-       // validateLike(filmLikesList, userId);
+        String sqlCreateLike = "INSERT INTO likes(film_id, user_id) VALUES (?, ?)";
+        jdbcTemplate.update(sqlCreateLike, filmId, userId);
 
-        //Запись like в DB
-        String sqlCreateLike =
-                "INSERT INTO likes(" +
-                        "FILM_ID, " +
-                        "USER_ID) " +
-                        "VALUES (?, ?)";
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update(connection -> {
-            PreparedStatement stmt = connection.prepareStatement(sqlCreateLike, new String[]{"LIKE_ID"});
-            stmt.setLong(1, filmId);
-            stmt.setLong(2, userId);
-            return stmt;
-        }, keyHolder);
-        Integer likeId = keyHolder.getKey().intValue();
         return getFilmById(filmId);
     }
 
 
     @Override
     public void deleteLike(Integer filmId, Integer userId) {
-        //Валидация user_id
-        String sqlFilmLikesById = "SELECT * FROM likes WHERE FILM_ID = ?";
-        List<Long> filmLikesList = jdbcTemplate.query(sqlFilmLikesById, (rs, rowNum) -> rs.getLong("USER_ID"), filmId);
-        //validateExist(filmLikesList, userId);
-
-        //Удаление записи like в DB
+        String sqlFilmLikesById = "SELECT * FROM likes WHERE film_id = ?";
+        List<Integer> likesList = jdbcTemplate.query(sqlFilmLikesById, (rs, rowNum) -> rs.getInt(
+                "user_id"), filmId);
+        if (!likesList.contains(userId)) {
+            throw new NotFoundException(String.format("Film with %s is not found", userId));
+        }
         String sqlDeleteLike = "DELETE FROM likes " +
-                "WHERE FILM_ID = ? AND USER_ID = ?";
+                "WHERE film_id = ? AND user_id = ?";
         jdbcTemplate.update(sqlDeleteLike, filmId, userId);
     }
+
 
     @Override
     public Collection<Film> getPopularFilmCount(int count) {
@@ -222,40 +216,21 @@ validateExistGenre(genresById, id);
             return popular;
         } else {
             return popular.stream()
-                    .sorted((film1, film2) -> {
-                        int result =
-                                Integer.compare(
-                                        getFilmLikes(film1.getId()).size()
-                                        , getFilmLikes(film2.getId()).size());
-                        result = -1 * result;
-                        return result;
-                    })
+                    .sorted((film1, film2) ->
+                            Integer.compare(
+                                    getFilmLikes(film2.getId()).size(),
+                                    getFilmLikes(film1.getId()).size())
+                    )
                     .limit(count)
                     .collect(Collectors.toList());
         }
     }
-    public List<Integer> getFilmLikes(Integer id) {
-        //Возврат likes
-        String sqlFilmLikesById = "SELECT * FROM likes WHERE FILM_ID = ?";
-        List<Integer> filmLikesList = jdbcTemplate.query(sqlFilmLikesById, (rs, rowNum) -> rs.getInt("USER_ID"), id);
-        return filmLikesList;
-    }
-//    public Collection<Film> getpularFilmCount(int count) {
-//        List<Film> popular = new ArrayList<>(getFilms());
-//        return getTopNFilms(popular, count);
-//    }
-//
-//    private List<Film> getTopNFilms(List<Film> films, int n) {
-//        Comparator<Film> comparator = new Comparator<Film>() {
-//            @Override
-//            public int compare(Film film1, Film film2) {
-//                return film2.getLikes().size() - film1.getLikes().size();
-//            }
-//        };
-//        Collections.sort(films, comparator);
-//        return null;//films.subList(0, Math.min(n, films.size()));
-//    }
 
+
+    public List<Integer> getFilmLikes(Integer id) {
+        String sqlFilmLikesById = "SELECT * FROM likes WHERE FILM_ID = ?";
+        return jdbcTemplate.query(sqlFilmLikesById, (rs, rowNum) -> rs.getInt("USER_ID"), id);
+    }
 
     static MPA makeMpa(ResultSet rs, int rowNum) throws SQLException {
         return new MPA(
@@ -283,30 +258,6 @@ validateExistGenre(genresById, id);
         rating.setName(rs.getString("rating_name"));
         film.setMpa(rating);
         return film;
-    }
-
-    public static void validateMPA(List<Integer> mpaIdList, Integer id) throws ValidationException {
-        if (!mpaIdList.contains(id)) {
-            throw new ValidationException(String.format("MPA with %s is not found", id));
-        }
-    }
-
-    public static void validateExistMPA(List<MPA> mpaIdList, Integer id) throws NotFoundException {
-        if (mpaIdList.isEmpty() || mpaIdList.get(0) == null || !mpaIdList.get(0).getId().equals(id)) {
-            throw new NotFoundException(String.format("MPA with %s is not found", id));
-        }
-    }
-
-    public static void validateGenre(List<Integer> genreList, Integer id) throws ValidationException {
-        if (!genreList.contains(id)) {
-            throw new ValidationException(String.format("Genre with %s is not found", id));
-        }
-    }
-
-    public static void validateExistGenre(List<Genre> genreList, Integer id) throws NotFoundException {
-        if (genreList.isEmpty() || genreList.get(0) == null || !genreList.get(0).getId().equals(id)) {
-            throw new NotFoundException(String.format("Genre with %s is not found", id));
-        }
     }
 }
